refactor: split contest label generation into per-type files
Extract label generators from monolithic contest.ts into:
- labels/axc.ts (ABC/ARC/AGC/ATC regex labels)
- labels/universities.ts (AtCoder university contest labels)
- labels/world_tour_finals.ts (WTF/AWTF labels)
- labels/atcoder_others.ts (OTHERS fallback chain)
- labels/joi.ts (JOI contest labels)
- labels/past.ts (PAST contest labels + translations)
- labels/aoj.ts (AOJ contest labels + translations)
- labels/index.ts (LABEL_GENERATORS Map + getContestNameLabel)
